requirements.txt
channels
channels-redis
django-markdownify
Pillow